requirements.txt
asgiref==3.4.1
black==22.8.0
click==8.0.4
dataclasses==0.8
Django==3.2.16
djangorestframework==3.14.0
flake8==5.0.4
importlib-metadata==4.2.0
isort==5.10.1
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
pathspec==0.9.0
pkg-resources==0.0.0
platformdirs==2.4.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.7
sqlparse==0.4.3
tomli==1.2.3
typed-ast==1.5.4
typing-extensions==4.1.1
zipp==3.6.0
